Add handling of blank columns and improve logic
Today's testing revealed that the program was failing if there wasn't
any information about a given column. Also, the logic for inclusions
was too lax. Words were getting through as long as they included one
of the known inclusions, but we only want words that have all of them.

The logic for exclusions was working, but it made sense to update it
alongside the inclusions logic to keep them consistent. I think the
code is a little more readable written this way anyway.